updated popularity script to use hacker news popularity algorithm
this commit changes the schema of solr which means that the solr needs
to be rebuilt when this is deployed.

+++ apparelrow/apparel/models.py
@@ -148,7 +148,7 @@ class Product(models.Model):
     gender        = models.CharField(_('Gender'), max_length=1, choices=PRODUCT_GENDERS, null=True, blank=True, db_index=True)
     feed_gender   = models.CharField(_('Feed gender'), max_length=1, choices=PRODUCT_GENDERS, null=True, blank=True, db_index=True)
     published     = models.BooleanField(default=True, db_index=True)
-    popularity    = models.IntegerField(default=0)
+    popularity    = models.DecimalField(default=0, max_digits=20, decimal_places=8)
     
     objects = SearchManager()
